+++ NEWS.md
@@ -1,7 +1,13 @@
-# chevron 0.2.0.9001
+# chevron 0.2.0.9017
 
 * Update `LBT06` template to use `PARAM` as row split.
 * Convert `AVISIT` to factor according to `AVISITN` as order in preprocessing.
+* Add `CFBT01` template. `VST01`, `EGT01` and `LBT01` are now all following `CFBT01`. By default parameters are by page displayed.
+* Add `RMPT04` and `RMPT05` which follow `RMPT01`.
+* Add `RMPT03` template.
+* Update `EGT03` to use `ACTARMCD` as default arm variable, and remove the preprocessing of filtering to "HR".
+* Add `COXT01` template. `COXT02` is based on `COXT01` now.
+* Add `AET05` and `AET05_ALL` template.
 
 # chevron 0.2.0
 
@@ -10,7 +16,7 @@
 * Remove deprecated `getter` functions `get_main`, `get_preprocess` and `get_postprocess`.
 * Simplify `pre` function and add more data checks in `main` function.
 * Remove redundant assertion functions.
-* Add more templates: `AET10`, `KMG01`, `RSPT01`, `RMPT01`, `COXT02`, `FSTG01`, `LBT06`.
+* Add more templates: `AET10`, `KMG01`, `RSPT01`, `RMPT01`, `COXT02`, `FSTG01`, and `LBT06`.
 
 # chevron 0.1.4